+++ assistants/core/src/rules/groups-max-layers/index.ts
@@ -18,16 +18,16 @@ export const createRule: CreateRuleFunction = (i18n) => {
       if (group._class === 'shapeGroup') continue // Do not consider shape groups, its common/expected for these to have many layers
       const numLayers = group.layers.filter((layer) => !skipClasses.includes(layer._class)).length
       if (numLayers > maxLayers) {
-        utils.report({
-          object: group,
-          message: i18n._(
+        utils.report(
+          i18n._(
             plural({
               value: numLayers,
               one: 'There is one layer in this group',
               other: 'There are # layers in this group',
             }),
           ),
-        })
+          group,
+        )
       }
     }
   }