+++ assistants/core/src/rules/shadows-no-disabled/index.ts
@@ -17,19 +17,13 @@ export const createRule: CreateRuleFunction = (i18n) => {
       if (!layer.style) continue // Narrow type to truthy `style` prop
       if (typeof layer.sharedStyleID === 'string') continue // Ignore layers using a shared style
       if (styleHasDisabledShadow(layer.style)) {
-        utils.report({
-          object: layer,
-          message: i18n._(t`There's a disabled shadow on this Layer Style`),
-        })
+        utils.report(i18n._(t`There's a disabled shadow on this Layer Style`), layer)
       }
     }
 
     for (const sharedStyle of utils.objects.sharedStyle) {
       if (styleHasDisabledShadow(sharedStyle.value)) {
-        utils.report({
-          object: sharedStyle,
-          message: i18n._(t`There's a disabled shadow on this Shared Style`),
-        })
+        utils.report(i18n._(t`There's a disabled shadow on this Shared Style`), sharedStyle)
       }
     }
   }