+++ packages/cli/src/index.ts
@@ -293,7 +293,16 @@ const formatResults = (cliResults: CliResults): string => {
         continue
       }
       append(2, `${chalk.inverse(` ${name} `)}\n`)
-      append(4, `Grade: ${res.result.passed ? chalk.green('pass') : chalk.red('fail')}`)
+      append(
+        4,
+        `Grade: ${
+          res.result.grade === 'pass'
+            ? chalk.green('pass')
+            : res.result.grade === 'fail'
+            ? chalk.red('fail')
+            : chalk.gray('unknown')
+        }`,
+      )
       append(4, `Violations: ${res.result.violations.length}`)
       append(4, `Rule errors: ${res.result.ruleErrors.length}\n`)
 
@@ -303,9 +312,14 @@ const formatResults = (cliResults: CliResults): string => {
         append(6, `Detail: ${chalk.grey(violation.message)}`)
         append(6, `Severity: ${chalk.grey(ViolationSeverity[violation.severity])}`)
         append(6, `Rule: ${chalk.grey(violation.ruleName)}`)
-        if (violation.objectName) append(6, `Layer: ${chalk.grey(violation.objectName)}`)
-        if (violation.pointer) append(6, `Location: ${chalk.grey(violation.pointer)}`)
-        if (violation.objectId) append(6, `Object: ${chalk.grey(violation.objectId)}`)
+        for (const object of violation.objects) {
+          const objects = []
+          if (object.name) objects.push(`Name : ${object.name}`)
+          if (object.pointer) objects.push(`Pointer : ${object.pointer}`)
+          if (object.id) objects.push(`Id : ${object.id}`)
+          if (object.class) objects.push(`Class : ${object.id}`)
+          append(6, `Object: ${chalk.grey(objects.join(' | '))}`)
+        }
         append()
       }
 
@@ -354,7 +368,13 @@ const runFile = async (filepath: string, tmpDir: string): Promise<RunOutput> => 
   spinner.start(spinnerMessage(filename, 'Processing file…'))
 
   try {
-    const operation = { cancelled: false }
+    const cancelToken = { cancelled: false }
+    const timeBudgets = {
+      totalMs: Infinity,
+      minRuleTimeoutMs: 0,
+      maxRuleTimeoutMs: Infinity,
+    }
+
     let file = await fromFile(filepath)
     const workspace = await getWorkspace(file)
     let { ignore = { pages: [], assistants: {} } } = workspace
@@ -362,7 +382,7 @@ const runFile = async (filepath: string, tmpDir: string): Promise<RunOutput> => 
     ignore = prunePages(ignore, file)
     file = filterPages(file, ignore.pages)
 
-    const processedFile = await processFile(file, operation)
+    const processedFile = await processFile(file, cancelToken)
     const env = {
       runtime: AssistantRuntime.Node,
       locale: await osLocale(),
@@ -391,8 +411,9 @@ const runFile = async (filepath: string, tmpDir: string): Promise<RunOutput> => 
       assistants,
       processedFile,
       getImageMetadata,
-      operation,
+      cancelToken,
       env,
+      timeBudgets,
     })
 
     spinner.succeed(spinnerMessage(filename, 'Ready'))